+++ kyo-net/jvm-native/src/main/scala/kyo/net/internal/posix/PosixTransport.scala
@@ -319,7 +319,7 @@ final private[net] class PosixTransport private[posix] (
                 promise.completeDiscard(Result.fail(connectFail(host, port, "")))
             case Present((family, addr, len)) =>
                 val sockR = sockets.socket(family, PosixConstants.SOCK_STREAM, 0)
-                val fd    = sockR.value.toInt
+                val fd    = sockR.value
                 if fd < 0 then
                     addr.close()
                     promise.completeDiscard(Result.fail(connectFail(host, port, new NetErrno(sockR.errorCode))))
@@ -702,7 +702,7 @@ final private[net] class PosixTransport private[posix] (
             case Present((family, addr, len)) =>
                 try
                     val sockR = sockets.socket(family, PosixConstants.SOCK_STREAM, 0)
-                    val fd    = sockR.value.toInt
+                    val fd    = sockR.value
                     if fd < 0 then
                         promise.completeDiscard(Result.fail(NetBindException(host, port, new NetErrno(sockR.errorCode))))
                     else
@@ -819,7 +819,7 @@ final private[net] class PosixTransport private[posix] (
                 @scala.annotation.tailrec
                 def drain(transientRetries: Int): AcceptDrain =
                     val r  = sockets.acceptNow(listener.serverFd, noAddr, noLen)
-                    val fd = r.value.toInt
+                    val fd = r.value
                     if fd >= 0 then
                         handleAccepted(fd, handler, tls)
                         drain(transientRetries)
